requirements.txt
# ── Core inference (local model) ──────────────────────────────────────────────
torch>=2.5.0
transformers>=4.47.0
accelerate>=0.34.0
bitsandbytes>=0.43.0      # 4-bit / 8-bit quantization

# ── Inference server ──────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
pydantic>=2.7.0

# ── HTTP client (agent ↔ local server) ───────────────────────────────────────
httpx>=0.27.0

# ── Model download ────────────────────────────────────────────────────────────
huggingface_hub>=0.23.0

# ── Utilities ─────────────────────────────────────────────────────────────────
rich>=13.7.0
typer>=0.12.0
pyyaml>=6.0
